requirements.txt
pyserial
pyserial-asyncio
autobahn[asyncio]